final review: narrow pressure_drops, scrub stale skip wording, add invariant test
Adversarial pre-merge review of the rollback (b08fd42).  Four
concrete fixes; nothing structural.

1. Narrow pressure_drops uint64_t → uint32_t
   ------------------------------------------
   The HTTP thread reads `output->pressure_drops` from
   query_transport_status while the producer thread writes it from
   *_observe_pressure.  uint64_t loads are not atomic on ARMv7 — the
   reader could see a torn value.  uint32_t fits the wire field
   (RtpSidecarTransportInfoWire.pressure_drops is already uint32_t)
   and is single-load atomic at our 4-byte alignment.  Wraps in 414
   minutes at 120 fps continuous pressure → 24855 days; fine for
   telemetry.

2. Stale "skip" / "drop frame" wording cleared
   -------------------------------------------
   Four call sites still implied the producer skips frames on
   pressure (they don't, since b08fd42):
     - include/venc_config.h:97   `outgoing.backpressure` doc said
       "drop entire frames", "Currently meaningful for shm://, unix://
       planned, udp:// out of scope".  Both halves stale — observation
       only, all three transports wired.  Rewrote to describe the
       telemetry-only path and the link_controller hand-off.
     - include/rtp_sidecar.h:150  trailer doc "frames the producer
       chose to skip" → "frames observed in pressure (ABI name
       retained pre-rollback)".
     - include/star6e_output.h, include/maruko_output.h field docs.
     - src/venc_api.c:661 validator comment: missing udp:// in the
       transport list.

3. Always-send invariant test
   --------------------------
   The b08fd42 rollback's load-bearing guarantee — "the producer
   never bails out of the send path on the basis of in_pressure" —
   had only indirect coverage via the hysteresis test.  New
   test_star6e_output_always_sends_under_pressure() drives the SHM
   ring above high_water, asserts the flag flips, then calls
   star6e_output_send_rtp_parts and checks ring->stats.writes
   actually advanced.  If a future change reintroduces a skip
   shortcut anywhere in the send path, this fails immediately.

4. Test label rename
   -----------------
   tests/test_venc_api.c CHECK strings still said "shm watermarks…",
   "shm bp on/off", "shm high>100", "shm lo>=hi", "shm multi…".
   Renamed to transport-neutral ("watermarks default…", "bp on/off",
   "high>100", "lo>=hi", "multi…") since the watermarks apply to
   shm/unix/udp uniformly now.

make verify clean both backends; tests 1404 → 1411 (+7 from new
always-send test).

+++ include/maruko_output.h
@@ -61,13 +61,12 @@ typedef struct {
 	uint32_t transport_gen; /* seqlock: odd = write in progress, even = stable */
 	int send_buf_capacity; /* cached SO_SNDBUF (kernel-reported), 0 = unknown */
 	MarukoOutputBatch batch;
-	/* Transport backpressure state.  Meaningful for any active transport
-	 * (ring fill for shm://, SIOCOUTQ / SO_SNDBUF for unix:// / udp://).
-	 * Hysteresis on cfg->outgoing.high_water_pct / low_water_pct;
-	 * pressure_drops surfaced via /api/v1/transport/status and the
-	 * rtp_sidecar transport trailer. */
+	/* Transport-pressure observation (telemetry only — never gates
+	 * frame transmission).  Mirrors Star6eOutput: hysteresis flag plus
+	 * a "frames observed in pressure" counter.  ABI-named pressure_drops
+	 * is uint32_t to keep the HTTP-thread read atomic on ARMv7. */
 	int in_pressure;
-	uint64_t pressure_drops;
+	uint32_t pressure_drops;
 } MarukoOutput;
 
 /** Initialize UDP or Unix socket output from a parsed URI. */
